requirements.txt
Django==4.2
djangorestframework==3.14.0
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
drf-spectacular==0.27.0
python-decouple==3.8
